Moved listToEither to ErrorHandling
